validation and simulation tests consolidated
